fix: 버전 컴포넌트 수정
