+++ src/pages/popup/components/VersionUpdateBar/index.ts
