+++ vite.config.ts
@@ -45,6 +45,6 @@ export default defineConfig({
     },
     define: {
         APP_VERSION: JSON.stringify(pkg.version),
-        ASSET_NAME: 'unofficial-flex-extension.zip',
+        ASSET_NAME: JSON.stringify('unofficial-flex-extension.zip'),
     },
 })
